Added poetry, pre-commit and other refactor
