--- nnmodel/layers/__init__.py
@@ -1,16 +1,0 @@
-from nnmodel.layers.dense import Dense
-from nnmodel.layers.conv2d import Conv2D
-from nnmodel.layers.conv2d_transpose import Conv2DTranspose
-from nnmodel.layers.rnn import RNN
-from nnmodel.layers.lstm import LSTM
-from nnmodel.layers.gru import GRU
-from nnmodel.layers.time_distributed import TimeDistributed
-from nnmodel.layers.batch_norm import BatchNormalization
-from nnmodel.layers.pooling2d import MaxPooling2D, AveragePooling2D
-from nnmodel.layers.upsampling2d import UpSampling2D
-from nnmodel.layers.activation import Activation
-from nnmodel.layers.dropout import Dropout
-from nnmodel.layers.flatten import Flatten
-from nnmodel.layers.zero_padding2d import ZeroPadding2D
-from nnmodel.layers.reshape import Reshape
-from nnmodel.layers.repeat_vector import RepeatVector
