+++ requirements.txt
@@ -1,0 +1,90 @@
+colorama==0.4.5 ; python_version >= "3.8" and python_version < "3.11" and platform_system == "Windows" \
+    --hash=sha256:854bf444933e37f5824ae7bfc1e98d5bce2ebe4160d46b5edf346a89358e99da \
+    --hash=sha256:e6c6b4334fc50988a639d9b98aa429a0b57da6e17b9a44f0451f930b6967b7a4
+llvmlite==0.38.1 ; python_version >= "3.8" and python_version < "3.11" \
+    --hash=sha256:0622a86301fcf81cc50d7ed5b4bebe992c030580d413a8443b328ed4f4d82561 \
+    --hash=sha256:0c0adce1793d66d009c554809f27baeb6258bf13f6fbaa12eff7443500caec25 \
+    --hash=sha256:0e609f7312a439b53b6f622d99180c3ff6a3e1e4ceca4d18aca1c5b46f4e3664 \
+    --hash=sha256:3d76c0fa42390bef56979ed213fbf0150c3fef36f5ea68d3d780d5d725da8c01 \
+    --hash=sha256:41e638a71c85a9a4a33f279c4cd812bc2f84122505b1f6ab8984ec7debb8548b \
+    --hash=sha256:4c1e5805c92e049b4956ed01204c6647de6160ab9aefb0d67ea83ca02a1d889a \
+    --hash=sha256:4e11bd9929dcbd55d5eb5cd7b08bf71b0097ea48cc192b69d102a90dd6e9816f \
+    --hash=sha256:5559e46c79b4017c3c25edc3b9512d11adc3689b9046120c685b0905c08d48a5 \
+    --hash=sha256:5c07d63df4578f31b39b764d3b4291f70157af7f42e171a8884ae7aaf989d1f7 \
+    --hash=sha256:633c9026eb43b9903cc4ffbc1c7d5293b2e3ad95d06fa9eab0f6ce6ff6ea15b3 \
+    --hash=sha256:66462d768c30d5f648ca3361d657b434efa8b09f6cf04d6b6eae66e62e993644 \
+    --hash=sha256:7a5e0ed215a576f0f872f47a70b8cb49864e0aefc8586aff5ce83e3bff47bc23 \
+    --hash=sha256:7db018da2863034ad9c73c946625637f3a89635bc70576068bab4bd085eea90d \
+    --hash=sha256:84d5a0163c172db2b2ae561d2fc0866fbd9f716cf13f92c0d41ca4338e682672 \
+    --hash=sha256:8c4f26c6c370e134a909ac555a671fa1376e74c69af0208f25c0979472577a9d \
+    --hash=sha256:8c64c90a8b0b7b7e1ed1912ba82c1a3f43cf25affbe06aa3c56c84050edee8ac \
+    --hash=sha256:9c8fac4edbadefa4dddf5dc6cca76bc2ae81df211dcd16a6638d60cc41249e56 \
+    --hash=sha256:9f53c3448410cc84d0e1af84dbc0d60ad32779853d40bcc8b1ee3c67ebbe94b1 \
+    --hash=sha256:a263252a68d85450110ec1f2b406c0414e49b04a4d216d31c0515ea1d59c3882 \
+    --hash=sha256:a7dd2bd1d6406e7789273e3f8a304ed5d9adcfaa5768052fca7dc233a857be98 \
+    --hash=sha256:ab070266f0f51304789a6c20d4be91a9e69683ad9bd4861eb89980e8eb613b3a \
+    --hash=sha256:b98da8436dbc29013ea301f1fdb0d596ab53bf0ab65c976d96d00bb6faa0b479 \
+    --hash=sha256:de8bd61480173930f2a029673e7cd0738fbbb5171dfe490340839ad7301d4cf0 \
+    --hash=sha256:ed7528b8b85de930b76407e44b080e4f376b7a007c2879749599ff8e2fe32753 \
+    --hash=sha256:edfa2c761cfa56cf76e783290d82e117f829bb691d8d90aa375505204888abac \
+    --hash=sha256:ef9aa574eff2e15f8c47b255da0db5dab326dc7f76384c307ae35490e2d2489a \
+    --hash=sha256:f95f455697c44d7c04ef95fdfce04629f48df08a832d0a0d9eb2363186dbb969 \
+    --hash=sha256:fbfbe546394c39db39a6898a51972aa131c8d6b0628517728b350552f58bdc19
+numba==0.55.2 ; python_version >= "3.8" and python_version < "3.11" \
+    --hash=sha256:09ff4d690abb05ffbb8a29a96d1cf35b46887a26796d3670de104beeec73d639 \
+    --hash=sha256:1105449247f338e49d63eb04a4aaa5c440bb5435df00f718c8e6e7afad841bb0 \
+    --hash=sha256:135fb7694928f9f57b4ff5b1be58f20f4771fedd1680636a9affdead96051959 \
+    --hash=sha256:16a52a0641c342b09b39f6762dcbe3846e44aa9baaaf4703b2ca42a3aee7346f \
+    --hash=sha256:20de0139d2267c8f0e2470d4f88540446cd1bf40de0f29f31b7ab9bf25d49b45 \
+    --hash=sha256:25410557d0deb1d97397b71e142a36772133986a7dd4fe2935786e2dd149245f \
+    --hash=sha256:2665ef28e900b3a55bf370daa81c12ebc64cd434116accd60c38a95a159a3182 \
+    --hash=sha256:29b89a68af162acf87adeb8fbf01f6bb1effae4711b28146f95108d82e905624 \
+    --hash=sha256:32649584144c35ced239937ab2c416ab22bbc1490ef8d90609c30fff9f6aa1b8 \
+    --hash=sha256:3eaf53e73e700370163e58257257299ac0d46fea4f244bf5476e4635bc31d808 \
+    --hash=sha256:46715180f87d5a1f3e4077d207ade66c96fc01159f5b7d49cee2d6ffb9e6539f \
+    --hash=sha256:5559c6684bf6cce7a22c656d8fef3e7c38ff5fec5153abef5955f6f7cae9f102 \
+    --hash=sha256:676c081162cc9403706071c1d1d42e479c0741551ab28096ba13859a2e3e9b80 \
+    --hash=sha256:68bb33eaef1d6155fc1ae4fa6c915b8a42e5052c89a58742254eaad072eab118 \
+    --hash=sha256:69b2e823efa40d32b259f5c094476dde2226b92032f17015d8cd7c10472654ce \
+    --hash=sha256:6e0f9b5d1c8ea1bdef39b0ad921a9bbf0cc4a88e76d722d756c68f1653787c35 \
+    --hash=sha256:8d5760a1e6a48d98d6b9cf774e4d2a64813d981cca60d7b7356af61195a6ca17 \
+    --hash=sha256:a669212aa66ffee4ad778016ac3819add33f9bcb96b4c384d3099531dd175085 \
+    --hash=sha256:a85779adc5234f7857615d1bd2c7b514314521f9f0163c33017707ed9816e6e6 \
+    --hash=sha256:d1c3cef3289fefb5673ceae32024ab5a8a08d4f4380bcb8348d01f1ba570ccff \
+    --hash=sha256:d7ac9ea5feef9536ab8bfbbb3ded1a0617ea8794d7547800d535b7857800f996 \
+    --hash=sha256:da4485e0f0b9562f39c78887149b33d13d787aa696553c9257b95575122905ed \
+    --hash=sha256:dcde1a1a3a430fb5f83c7e095b0b6ac7adb5595f50a3ee05babb2964f31613c4 \
+    --hash=sha256:dd05f7c0ce64b6977596aa4e5a44747c6ef414d7989da1c7672337c54381a5ef \
+    --hash=sha256:de1f93bd7e2d431451aec20a52ac651a020e98a4ba46797fad860bba338a7e64 \
+    --hash=sha256:dfddd633141608a09cbce275fb9fe7aa514918625ace20b0e587898a2d93c030 \
+    --hash=sha256:e36232eccd172c583b1f021c5c48744c087ae6fc9dc5c5f0dd2cb2286e517bf8 \
+    --hash=sha256:e428d9e11d9ba592849ccc9f7a009003eb7d30612007e365afe743ce7118c6f4
+numpy==1.22.4 ; python_version >= "3.8" and python_version < "3.11" \
+    --hash=sha256:0791fbd1e43bf74b3502133207e378901272f3c156c4df4954cad833b1380207 \
+    --hash=sha256:1ce7ab2053e36c0a71e7a13a7475bd3b1f54750b4b433adc96313e127b870887 \
+    --hash=sha256:2d487e06ecbf1dc2f18e7efce82ded4f705f4bd0cd02677ffccfb39e5c284c7e \
+    --hash=sha256:37431a77ceb9307c28382c9773da9f306435135fae6b80b62a11c53cfedd8802 \
+    --hash=sha256:3e1ffa4748168e1cc8d3cde93f006fe92b5421396221a02f2274aab6ac83b077 \
+    --hash=sha256:425b390e4619f58d8526b3dcf656dde069133ae5c240229821f01b5f44ea07af \
+    --hash=sha256:43a8ca7391b626b4c4fe20aefe79fec683279e31e7c79716863b4b25021e0e74 \
+    --hash=sha256:4c6036521f11a731ce0648f10c18ae66d7143865f19f7299943c985cdc95afb5 \
+    --hash=sha256:59d55e634968b8f77d3fd674a3cf0b96e85147cd6556ec64ade018f27e9479e1 \
+    --hash=sha256:64f56fc53a2d18b1924abd15745e30d82a5782b2cab3429aceecc6875bd5add0 \
+    --hash=sha256:7228ad13744f63575b3a972d7ee4fd61815b2879998e70930d4ccf9ec721dce0 \
+    --hash=sha256:9ce7df0abeabe7fbd8ccbf343dc0db72f68549856b863ae3dd580255d009648e \
+    --hash=sha256:a911e317e8c826ea632205e63ed8507e0dc877dcdc49744584dfc363df9ca08c \
+    --hash=sha256:b89bf9b94b3d624e7bb480344e91f68c1c6c75f026ed6755955117de00917a7c \
+    --hash=sha256:ba9ead61dfb5d971d77b6c131a9dbee62294a932bf6a356e48c75ae684e635b3 \
+    --hash=sha256:c1d937820db6e43bec43e8d016b9b3165dcb42892ea9f106c70fb13d430ffe72 \
+    --hash=sha256:cc7f00008eb7d3f2489fca6f334ec19ca63e31371be28fd5dad955b16ec285bd \
+    --hash=sha256:d4c5d5eb2ec8da0b4f50c9a843393971f31f1d60be87e0fb0917a49133d257d6 \
+    --hash=sha256:e96d7f3096a36c8754207ab89d4b3282ba7b49ea140e4973591852c77d09eb76 \
+    --hash=sha256:f0725df166cf4785c0bc4cbfb320203182b1ecd30fee6e541c8752a92df6aa32 \
+    --hash=sha256:f3eb268dbd5cfaffd9448113539e44e2dd1c5ca9ce25576f7c04a5453edc26fa \
+    --hash=sha256:fb7a980c81dd932381f8228a426df8aeb70d59bbcda2af075b627bbc50207cba
+setuptools==62.6.0 ; python_version >= "3.8" and python_version < "3.11" \
+    --hash=sha256:990a4f7861b31532871ab72331e755b5f14efbe52d336ea7f6118144dd478741 \
+    --hash=sha256:c1848f654aea2e3526d17fc3ce6aeaa5e7e24e66e645b5be2171f3f6b4e5a178
+tqdm==4.64.0 ; python_version >= "3.8" and python_version < "3.11" \
+    --hash=sha256:40be55d30e200777a307a7585aee69e4eabb46b4ec6a4b4a5f2d9f11e7d5408d \
+    --hash=sha256:74a2cdefe14d11442cedf3ba4e21a3b84ff9a2dbdc6cfae2c34addb2a14a5ea6
