+++ src/nnmodel/layers/gru.py
@@ -1,0 +1,14 @@
+import numpy as np
+
+
+class GRU:
+    def __init__(self):
+        pass
+
+    def forward_prop(self, X, training):
+        pass
+
+    def backward_prop(
+        self,
+    ):
+        pass
